+++ ui/src/components/card/QasCard.vue
@@ -6,6 +6,9 @@
           <q-carousel v-model="slideImage" animated class="cursor-pointer" height="205px" infinite :navigation="hasImages" navigation-icon="o_fiber_manual_record" swipeable>
             <q-carousel-slide v-for="(item, index) in imagesList" :key="index" class="bg-no-repeat" :class="bgImagePositionClasses" :img-src="item" :name="index" />
           </q-carousel>
+          <div class="absolute-top flex items-center q-pa-md">
+            <slot name="carousel-header" />
+          </div>
         </slot>
       </div>