+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Helper to translate plain string object to a complete object
 - Using helper on nested fields to translate errors
+- `QasNestedFields` enumerate labels.
+- Slot carousel header in component (`QasCard`). 
 
 ## 2.7.4 - 2021-08-25