+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased 
 
-### Fixed
+### Changed
 - Prop data-cy at `qas-form-view` changed to kebab-case.
 
 ## 2.8.0 - 2021-08-27