+++ CHANGELOG.md
@@ -9,6 +9,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Card image width size
 
+## 2.8.0 - 2021-08-27
+
+### Added
+- Props `readonly` to `QasSignatureUploader`.
+- Helper to translate plain string object to a complete object
+- Using helper on nested fields to translate errors
+- `QasNestedFields` enumerate labels.
+- Slot carousel header in component (`QasCard`). 
+
+## 2.7.4 - 2021-08-25
+
+### Fixed
+- Fixed `map-markers.js` icon import was breaking build on other applications due to non existing file.
+
+## 2.7.3 - 2021-08-24
+
+### Added
+- DataCy prop to `qas-field`.
+- DataCy prop to buttons at `qas-form-view`
+
+### Changed
+- Changed implementation of `destroy` on `QasDelete` component. Now it receives an `url prop` to override the destroy URL.
+- Changed implementation of `$_createMarkers` on `map-markers.js` mixin
+
 ## 2.7.2 - 2021-08-06
 
 ### Fixed